requirements.txt
aiofiles==24.1.0 ; python_version >= "3.12" and python_version < "4.0"
aiosmtplib==2.0.2 ; python_version >= "3.12" and python_version < "4.0"
alembic==1.13.2 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
argon2-cffi-bindings==21.2.0 ; python_version >= "3.12" and python_version < "4.0"
argon2-cffi==23.1.0 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.29.0 ; python_version >= "3.12" and python_version < "4.0"
bcrypt==4.1.3 ; python_version >= "3.12" and python_version < "4.0"
beanie-batteries-queue==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
beanie==1.26.0 ; python_version >= "3.12" and python_version < "4.0"
beanie[queue]==1.26.0 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.8.2 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.7.4 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
cryptography==42.0.8 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
ecdsa==0.19.0 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi-cli==0.0.4 ; python_version >= "3.12" and python_version < "4.0"
fastapi-mail==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
fastapi-sso==0.15.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.111.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi[all]==0.111.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
hiredis==2.3.2 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.7 ; python_version >= "3.12" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
lazy-model==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
mako==1.3.5 ; python_version >= "3.12" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
more-itertools==10.3.0 ; python_version >= "3.12" and python_version < "4.0"
motor==3.5.1 ; python_version >= "3.12" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.10.6 ; python_version >= "3.12" and python_version < "4.0"
passlib[argon2,bcrypt]==1.7.4 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.20.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-extra-types==2.9.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.3.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.8.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic[email]==2.8.2 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.12" and python_version < "4.0"
pymongo==4.8.0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.12" and python_version < "4.0"
python-ulid==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
redis-om==0.3.1 ; python_version >= "3.12" and python_version < "4.0"
redis==5.0.7 ; python_version >= "3.12" and python_version < "4.0"
rich==13.7.1 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
setuptools==69.5.1 ; python_version >= "3.12" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.31 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy[asyncio]==2.0.31 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.37.2 ; python_version >= "3.12" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.12" and python_version < "4.0"
typer==0.12.3 ; python_version >= "3.12" and python_version < "4.0"
types-cffi==1.16.0.20240331 ; python_version >= "3.12" and python_version < "4.0"
types-pyopenssl==24.1.0.20240425 ; python_version >= "3.12" and python_version < "4.0"
types-redis==4.6.0.20240425 ; python_version >= "3.12" and python_version < "4.0"
types-setuptools==70.3.0.20240710 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
ujson==5.10.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn[standard]==0.30.1 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.12" and python_version < "4.0"
watchfiles==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.12" and python_version < "4.0"
